Improve ModelPusher performance
- Move vanilla shading reversal to compute shader

+++ src/main/resources/rs117/hd/frag.glsl
@@ -39,7 +39,7 @@ uniform mat4 lightProjectionMatrix;
 uniform float elapsedTime;
 uniform float colorBlindnessIntensity;
 uniform vec3 fogColor;
-uniform int fogDepth;
+uniform float fogDepth;
 uniform vec3 waterColorLight;
 uniform vec3 waterColorMid;
 uniform vec3 waterColorDark;
